requirements/base.txt
djangorestframework==3.15.2
django-environ==0.11.2
django_cors_headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.0.0
pillow-avif-plugin==1.4.6
drf-spectacular==0.28.0
django-redis==5.4.0
colorlog==6.9.0
django-silk==5.3.2
python-dateutil==2.9.0.
